fix commands and queries for migrated workflows

+++ internal/app/project/delete_project_service.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/lubosgarancovsky/eden-inri/internal/app/ports"
 	"github.com/lubosgarancovsky/eden-inri/internal/domain/command"
+	"github.com/lubosgarancovsky/go-kit"
 )
 
 type DeleteProjectService struct {
@@ -18,5 +19,14 @@ func NewDeleteProjectService(repo ports.PersistProjectPort) *DeleteProjectServic
 }
 
 func (s *DeleteProjectService) Execute(ctx context.Context, cmd *command.DeleteCommand) error {
+	project, err := s.repo.FindByID(ctx, cmd.UserID, cmd.ID)
+	if err != nil {
+		return err
+	}
+
+	if !project.IsOwner() {
+		return go_kit.ErrForbidden.WithMessage("only owner can delete project")
+	}
+
 	return s.repo.Delete(ctx, cmd.ID)
 }
